Handle usernames with hyphens in comments
When extracting usernames from embedded comments using key phrases, allow
usernames that contain a hyphen to be extracted in full.  (Prior to this
change, only the part of the username before the hyphen would be extracted.

This commit also changes the regex used to allow at most one full stop or
hyphen in the user name, and escapes the special characters in the literals
being matched against.

Some basic unit tests for this part of the code have been added, including
for the new hyphen username case.  The tests relating to issue handling (which
are independent of the underlying source of the data) have been refactored
out into a separate test class/file to indicate this.