Convert GitLab issue state events into comments
A change to the GitLab API means that changes to issue state are
no longer recorded in issue comments, but as new issue state events.
Eventually, RoboTA Core should consider migrating its own model to
include such events, but for now we convert them into the old
comment model.

This requires a recent version of python-gitlab to pick up the new API.

This commit also removes the code that returns specific time recording
phrases.  This is now handled by RoboTA in code specific to the
coursework exercise that needs the phrases.